pyproject.toml
[project]
name = "wispr-clone"
version = "0.1.0"
description = "Personal Wispr-Flow-style dictation tool for Windows."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.11",
    "sounddevice>=0.4.6",
    "soundfile>=0.12",
    "numpy>=1.26",
    "keyboard>=0.13.5",
    "pyperclip>=1.8.2",
    "pystray>=0.19.5",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pyinstaller>=6.6"]

[project.scripts]
wispr-clone = "wispr_clone.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wispr_clone"]
